move the SIntendedSubmitInfo struct out of IUtilities
